osa4

+++ osa4/blogilista/controllers/blogs.js
@@ -1,7 +1,8 @@
 const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
-const User = require('../models/user')
+// const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -13,15 +14,19 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  console.log('tullaanko tänne asti 1')
+  
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  console.log('tullaanko tänne asti2')
+  
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
 
   
-  const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: 'content missing' })
@@ -44,7 +49,36 @@ blogsRouter.post('/', async (request, response) => {
 
 
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const user = request.user
+
+  
+  if (!user) {
+    return response.status(404).json({ error: 'user not found' })
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+  
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+ 
+
+  if (! blog.user._id.toString() === user.id.toString() ) {
+    return response.status(403).json({ error: 'only the user who added the blog can delete it' })
+  }
+  
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
